[Silver III] Title: 주유소, Time: 452 ms, Memory: 54976 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.kt
@@ -14,7 +14,7 @@ fun main() {
 
     for (i in distance.indices) {
         minCost = min(minCost, cost[i])
-        answer += (minCost * distance[i])
+        answer += (minCost.toLong() * distance[i].toLong())
     }
 
     bw.write("$answer")
